See [release notes](https://github.com/nautechsystems/nautilus_trader/releases/tag/v1.151.0).

+++ nautilus_core/persistence/src/filter_expr.rs
@@ -13,7 +13,6 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-use arrow2;
 use arrow2::array::UInt64Array;
 use arrow2::compute::comparison::primitive::eq_scalar;
 
@@ -21,5 +20,5 @@ use arrow2::compute::comparison::primitive::eq_scalar;
 fn main() {
     //WIP: Filter the arrays according to the filter_expr(s) argument.
     let bid = UInt64Array::from_vec(vec![1, 2, 3, 4, 5]);
-    let _boolean_mask = eq_scalar(&bid, 2 as u64);
+    let _boolean_mask = eq_scalar(&bid, 2_u64);
 }
